+++ system/platform-core/src/test/java/org/platformlambda/core/UtilityTests.java
@@ -64,6 +64,40 @@ class UtilityTests {
     }
 
     @Test
+    void deepCopyTest() {
+        var mm = new MultiLevelMap();
+        mm.setElement("hello.world", Map.of("a", List.of(10,20, Map.of("test", "message"))));
+        mm.setElement("hello.list[0]", List.of("x", Map.of("y", 200, "z", List.of(1,2,3))));
+        mm.setElement("hello.list-of-list",
+                        List.of(List.of(0,2,List.of(4,5,6)), "xyz", List.of("1", Map.of("k2", "v2"))));
+        var listWithNull = new ArrayList<>();
+        listWithNull.add(null);
+        listWithNull.add("12345");
+        mm.setElement("list.contains-null", listWithNull);
+        var mapWithNull = new HashMap<>();
+        mapWithNull.put("x", "y");
+        mapWithNull.put("nothing", null);
+        mm.setElement("map-with-null", mapWithNull);
+        Utility util = Utility.getInstance();
+        var result = util.deepCopy(mm.getMap());
+        var flat = util.getFlatMap(result);
+        for (var entry: flat.entrySet()) {
+            assertEquals(entry.getValue(), mm.getElement(entry.getKey()));
+        }
+        var copied = util.deepCopy(mm.getMap());
+        // map's null elements are not copied
+        assertNotEquals(mm.getMap(), copied);
+        mm.removeElement("map-with-null.nothing");
+        // should be the same after removing null element from a map
+        assertEquals(mm.getMap(), copied);
+        // modify the original map
+        mm.setElement("modified", true);
+        // prove that the original map and the cloned map are different
+        assertNotEquals(mm.getMap(), copied);
+        assertNotEquals(mm.getMap().size(), copied.size());
+    }
+
+    @Test
     void scanEnvVars() {
         Utility util = Utility.getInstance();
         var statement = "hello.${world}.this.is.a.${nice}.day";